# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SkillServe API",
        "description": "REST API for the SkillServe admin web system — a modular monolith (Laravel 13 + Sanctum). All responses use the standard envelope: { success, message, data, errors, meta }.",
        "contact": {
            "email": "group6@skillserve.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "List roles",
                "description": "GET /api/roles — paginated, searchable list of roles.",
                "operationId": "e4a0739c81c14d3ff5f3f17a0565628e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort column",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at",
                            "enum": [
                                "name",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Request successful.",
                                    "data": [
                                        {
                                            "id": 3,
                                            "name": "reports-manager",
                                            "description": "Manages operational reports.",
                                            "guard_name": "web",
                                            "permissions": [
                                                "view reports"
                                            ],
                                            "created_at": "2026-08-07T08:00:00+00:00"
                                        }
                                    ],
                                    "errors": null,
                                    "meta": {
                                        "pagination": {
                                            "total": 1,
                                            "per_page": 15,
                                            "current_page": 1,
                                            "last_page": 1,
                                            "from": 1,
                                            "to": 1
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated / expired token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Missing the manage administrators permission",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Create a role",
                "description": "POST /api/roles — create a new role.",
                "operationId": "da571ea2b2dc49eb6de0b4d1ee838448",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "reports-manager"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Manages operational reports."
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "name": "reports-manager",
                                "description": "Manages operational reports."
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Role created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Role created.",
                                    "data": {
                                        "id": 3,
                                        "name": "reports-manager",
                                        "description": "Manages operational reports.",
                                        "guard_name": "web",
                                        "permissions": [],
                                        "created_at": "2026-08-07T08:00:00+00:00"
                                    },
                                    "errors": null,
                                    "meta": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated / expired token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Missing the manage administrators permission",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/roles/{role}/permissions": {
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Sync role permissions",
                "description": "PUT /api/roles/{role}/permissions — sync the permission matrix for a role.",
                "operationId": "fd97d39d6940f05e370f157530a4d41e",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "permissions"
                                ],
                                "properties": {
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "view reports",
                                            "manage bookings"
                                        ]
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "permissions": [
                                    "view reports",
                                    "manage bookings"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role permissions synced",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Role permissions updated.",
                                    "data": {
                                        "id": 3,
                                        "name": "reports-manager",
                                        "description": "Manages operational reports.",
                                        "guard_name": "web",
                                        "permissions": [
                                            "view reports",
                                            "manage bookings"
                                        ],
                                        "created_at": "2026-08-07T08:00:00+00:00"
                                    },
                                    "errors": null,
                                    "meta": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated / expired token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Missing the manage administrators permission",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error / super administrator permissions immutable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiEnvelope": {
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Request successful."
                    },
                    "data": {
                        "description": "Payload (object/array/null)",
                        "nullable": true
                    },
                    "errors": {
                        "description": "Field errors keyed by field",
                        "type": "object",
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Paste the Sanctum token returned by POST /api/auth/login.",
                "bearerFormat": "sanctum-token",
                "scheme": "bearer"
            }
        }
    },
...
}